feat: Allow to delete Rule even not ended yet - MEED-1690 - Meeds-io/MIPs#49 (#1008)
- Prior to this change, for a challenge or automatic rule, the program
owner wasn't able to delete it. This change will allow to delete all
rules independently from its statuses.

- This change will fix creating announcement on deleted Rule and will deny
access to marked as deleted rules from UI.

- Prior to this change, a simple user can directly access a program
through direct URL. This change will ensure to secure program detail
screen access by checking user permissions in REST call.

+++ services/src/test/java/org/exoplatform/addons/gamification/utils/UtilsTest.java
@@ -174,7 +174,7 @@ public class UtilsTest extends AbstractServiceTest {
     RuleEntity ruleEntity = newRule("challenge1", domainEntity.getId());
     newRealizationEntityWithRuleId("annoucement 1", ruleEntity.getId());
     newRealizationEntityWithRuleId("annoucement 2", ruleEntity.getId());
-    assertEquals((Long) 2l, announcementService.countAnnouncements(ruleEntity.getId()));
+    assertEquals(2, announcementService.countAnnouncements(ruleEntity.getId()));
     assertThrows(ObjectNotFoundException.class, () -> announcementService.countAnnouncements(528l));
   }
